# apps/control-plane/src/interfaces/http/gateway/routes.ts
import {randomUUID} from 'node:crypto';
import type {FastifyInstance,FastifyRequest} from 'fastify';
import type {ModelGatewayCommand} from '../../../application/gateway/execute-model-request.js';

type Body=Readonly<{schema_version?:string;idempotency_key?:string;task?:string;messages?:readonly Readonly<{role:'system'|'user'|'assistant';content:string}>[];required_capabilities?:ModelGatewayCommand['requiredCapabilities'];maximum_output_tokens?:number;[key:string]:unknown}>;
export type ModelStreamEvent=Readonly<{schema_version:'1.0.0';request_id:string;sequence:number;type:'started'|'delta'|'completed'|'cancelled'|'failed';occurred_at:string;text?:string;code?:string;semantic_state?:string;citation_refs?:readonly string[]}>;
export type GatewayRouteServices=Readonly<{authorize(request:FastifyRequest):Promise<Pick<ModelGatewayCommand,'tenantId'|'workspaceId'|'userId'|'deviceId'|'sessionId'>>;execute(command:ModelGatewayCommand):Promise<Readonly<Record<string,unknown>>>;stream?(command:ModelGatewayCommand,signal:AbortSignal):AsyncIterable<ModelStreamEvent>;countInputTokens(messages:NonNullable<Body['messages']>):number;scopeKeys(context:Pick<ModelGatewayCommand,'tenantId'|'workspaceId'|'userId'>,task:string):ModelGatewayCommand['scopeKeys']}>;

const allowed=new Set(['schema_version','idempotency_key','task','messages','required_capabilities','maximum_output_tokens']);
function code(error:unknown){const value=String(error);for(const known of ['GATEWAY_REQUEST_INVALID','GATEWAY_REQUEST_IN_PROGRESS','POLICY_DENIED','ROUTE_DENIED','EGRESS_DENIED','BUDGET_DENIED','PROVIDER_UNAVAILABLE','GATEWAY_AUDIT_UNAVAILABLE'])if(value.includes(known))return known;return 'GATEWAY_DEPENDENCY_UNAVAILABLE';}
function status(value:string){if(value==='POLICY_DENIED')return 403;if(value==='GATEWAY_REQUEST_IN_PROGRESS')return 409;if(['ROUTE_DENIED','EGRESS_DENIED','BUDGET_DENIED'].includes(value))return 422;if(value==='PROVIDER_UNAVAILABLE')return 502;if(value==='GATEWAY_REQUEST_INVALID')return 400;return 503;}
function validate(body:Body){if(Object.keys(body).some(key=>!allowed.has(key))||body.schema_version!=='1.0.0'||!body.idempotency_key||!body.task||!body.messages?.length||!body.required_capabilities||!Number.isSafeInteger(body.maximum_output_tokens)||body.maximum_output_tokens!<=0)throw new Error('GATEWAY_REQUEST_INVALID');}
async function command(request:FastifyRequest,body:Body,services:GatewayRouteServices):Promise<ModelGatewayCommand>{validate(body);const actor=await services.authorize(request);return {...actor,requestId:randomUUID(),idempotencyKey:body.idempotency_key!,task:body.task!,messages:body.messages!,requiredCapabilities:body.required_capabilities!,inputTokens:services.countInputTokens(body.messages!),maximumOutputTokens:body.maximum_output_tokens!,scopeKeys:services.scopeKeys(actor,body.task!)};}
export function formatSseEvent(event:ModelStreamEvent){if(event.type==='completed'&&!event.semantic_state)throw new Error('GATEWAY_STREAM_SEMANTIC_STATE_REQUIRED');return `id: ${event.sequence}\nevent: ${event.type}\ndata: ${JSON.stringify(event)}\n\n`;}

export function registerGatewayRoutes(app:FastifyInstance,services:GatewayRouteServices){
 app.post<{Body:Body}>('/v1/model/responses',async(request,reply)=>{try{return reply.send(await services.execute(await command(request,request.body??{},services)));}catch(error){const safe=code(error);return reply.code(status(safe)).send({code:safe});}});
 app.post<{Body:Body}>('/v1/model/responses/stream',async(request,reply)=>{
  if(!String(request.headers.accept??'').split(',').some(value=>value.trim().startsWith('text/event-stream')))return reply.code(406).send({code:'GATEWAY_ACCEPT_INVALID'});
  try{
   if(!services.stream)throw new Error('GATEWAY_DEPENDENCY_UNAVAILABLE');
   const input=await command(request,request.body??{},services),controller=new AbortController();
   request.raw.once('close',()=>controller.abort());
   reply.hijack();reply.raw.writeHead(200,{'Content-Type':'text/event-stream; charset=utf-8','Cache-Control':'no-store','Connection':'keep-alive','X-Accel-Buffering':'no'});
   try{for await(const event of services.stream(input,controller.signal)){if(controller.signal.aborted)break;if(!reply.raw.write(formatSseEvent(event)))await new Promise<void>(resolve=>reply.raw.once('drain',resolve));}}finally{reply.raw.end();}
   return reply;
  }catch(error){if(reply.sent)return reply;const safe=code(error);return reply.code(status(safe)).send({code:safe});}
 });
}