# contracts/openapi/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Domus cross-runtime HTTP API", "version": "1.0.0", "description": "Interfaces internas; autenticação de workload é obrigatória no transporte." },
  "servers": [{ "url": "https://internal.invalid/v1" }],
  "paths": {
...
    "/model/responses/stream": { "post": { "operationId": "streamModelResponse", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "../../json-schema/v1/model-gateway-request.schema.json" } } } }, "responses": { "200": { "description": "SSE tipado, correlacionado e sem cache", "content": { "text/event-stream": { "schema": { "$ref": "../../json-schema/v1/model-stream-event.schema.json" } } } }, "default": { "$ref": "#/components/responses/ContractError" } } } }
  },
  "components": { "responses": { "ContractError": { "description": "Bloqueio tipado e fail-closed", "content": { "application/problem+json": { "schema": { "$ref": "../../json-schema/v1/contract-error.schema.json" } } } } } }
}